+++ packages/components/src/Dialog/Confirm/ConfirmLayout.tsx
@@ -27,7 +27,7 @@ import React, { FC, ReactElement } from 'react'
 import { ButtonProps } from '../../Button'
 import { IconProps } from '../../Icon'
 import { Paragraph } from '../../Text'
-import { ModalContent, ModalFooter, ModalHeader } from '../Layout'
+import { DialogContent, DialogFooter, DialogHeader } from '../Layout'
 
 interface ConfirmLayoutProps {
   /**
@@ -61,20 +61,20 @@ export const ConfirmLayout: FC<ConfirmLayoutProps> = ({
 }) => {
   return (
     <>
-      <ModalHeader hideClose headerIcon={titleIcon}>
+      <DialogHeader hideClose headerIcon={titleIcon}>
         {title}
-      </ModalHeader>
-      <ModalContent innerProps={{ py: 'none' }}>
+      </DialogHeader>
+      <DialogContent innerProps={{ py: 'none' }}>
         {typeof message === 'string' ? (
           <Paragraph breakword>{message}</Paragraph>
         ) : (
           message
         )}
-      </ModalContent>
-      <ModalFooter>
+      </DialogContent>
+      <DialogFooter>
         {primaryButton}
         {secondaryButton}
-      </ModalFooter>
+      </DialogFooter>
     </>
   )
 }