+++ packages/components/src/Tooltip/Tooltip.tsx
@@ -36,7 +36,7 @@ import React, {
   Ref,
 } from 'react'
 import omit from 'lodash/omit'
-import { ModalContext } from '../Dialog'
+import { DialogContext } from '../Dialog'
 import {
   useCallbackRef,
   useID,
@@ -203,7 +203,7 @@ export function useTooltip({
 
   const popper =
     isOpen && content && !disabled ? (
-      <ModalContext.Provider value={{ closeModal: handleClose }}>
+      <DialogContext.Provider value={{ closeModal: handleClose }}>
         <OverlaySurface
           arrow={arrow}
           arrowProps={arrowProps}
@@ -226,7 +226,7 @@ export function useTooltip({
             {content}
           </TooltipContent>
         </OverlaySurface>
-      </ModalContext.Provider>
+      </DialogContext.Provider>
     ) : null
 
   return {