+++ CHANGELOG.md
@@ -16,6 +16,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `Modal*` has moved to `Dialog*`
+  - `ModalContext` is now `DialogContext`
+  - `ModalContent` is now `DialogContent`
+  - `ModalFooter` is now `DialogFooter`
+  - `ModalHeader` is now `DialogHeader`
+  - Aliases for the old names are in place to ease the upgrade but these aliases will be deprecated shortly
 - `Badge` now uses generated colors rather than referencing colors.palette
 - `InputText` interface simplified / narrowed
   - No longer supports typography or pseudo props
@@ -38,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- `Drawer` no longer available
 - `ComboboxMultiOption` prop `hideCheckMark` (instead use `indicator={false}`)
 - `CustomizableAttributes` are no longer supported
 - `MenuList`, `MenuGroup` & `MenuItem` no longer supports `customizationProps`